+++ src/Unchase.Swashbuckle.AspNetCore.Extensions/Extensions/SwaggerGenOptionsExtensions.cs
@@ -1,5 +1,5 @@
-﻿using System;
-using System.ComponentModel;
+﻿using System.ComponentModel;
+using System.Net;
 using Microsoft.Extensions.DependencyInjection;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using Unchase.Swashbuckle.AspNetCore.Extensions.Filters;
@@ -8,22 +8,69 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Extensions
 {
     public static class SwaggerGenOptionsExtensions
     {
-        public static void AddEnumsWithValuesFixFilters(this SwaggerGenOptions swaggerGenOptions, bool includeDescriptionFromAttribute = false)
+        #region Extensions
+
+        /// <summary>
+        /// Change all responses by specific http status codes in OpenApi document.
+        /// </summary>
+        /// <typeparam name="T">Type of response example.</typeparam>
+        /// <param name="swaggerGenOptions"><see cref="SwaggerGenOptions"/>.</param>
+        /// <param name="httpStatusCode">HTTP status code.</param>
+        /// <param name="responseDescription">Response description.</param>
+        /// <param name="responseExampleOption"><see cref="ResponseExampleOptions"/>.</param>
+        /// <param name="responseExample">New example for response.</param>
+        /// <returns>
+        /// Returns <see cref="SwaggerGenOptions"/>.
+        /// </returns>
+        public static SwaggerGenOptions ChangeAllResponsesByHttpStatusCode<T>(
+            this SwaggerGenOptions swaggerGenOptions, 
+            int httpStatusCode,
+            string responseDescription = null,
+            ResponseExampleOptions responseExampleOption = ResponseExampleOptions.None,
+            T responseExample = default) where T : class
+        {
+            swaggerGenOptions.DocumentFilter<ChangeResponseByHttpStatusCodeDocumentFilter<T>>(httpStatusCode, responseDescription, responseExampleOption, responseExample);
+            return swaggerGenOptions;
+        }
+
+        /// <summary>
+        /// Change all responses by specific http status codes in OpenApi document.
+        /// </summary>
+        /// <typeparam name="T">Type of response example.</typeparam>
+        /// <param name="swaggerGenOptions"><see cref="SwaggerGenOptions"/>.</param>
+        /// <param name="httpStatusCode">HTTP status code.</param>
+        /// <param name="responseDescription">Response description.</param>
+        /// <param name="responseExampleOption"><see cref="ResponseExampleOptions"/>.</param>
+        /// <param name="responseExample">New example for response.</param>
+        /// <returns>
+        /// Returns <see cref="SwaggerGenOptions"/>.
+        /// </returns>
+        public static SwaggerGenOptions ChangeAllResponsesByHttpStatusCode<T>(
+            this SwaggerGenOptions swaggerGenOptions,
+            HttpStatusCode httpStatusCode,
+            string responseDescription = null,
+            ResponseExampleOptions responseExampleOption = ResponseExampleOptions.None,
+            T responseExample = default) where T : class
+        {
+            return swaggerGenOptions.ChangeAllResponsesByHttpStatusCode((int)httpStatusCode, responseDescription, responseExampleOption, responseExample);
+        }
+
+        /// <summary>
+        /// Add filters to fix enums in OpenApi document.
+        /// </summary>
+        /// <param name="swaggerGenOptions"><see cref="SwaggerGenOptions"/>.</param>
+        /// <param name="includeDescriptionFromAttribute">If true - add extensions (descriptions) from <see cref="DescriptionAttribute"/>.</param>
+        /// <returns>
+        /// Returns <see cref="SwaggerGenOptions"/>.
+        /// </returns>
+        public static SwaggerGenOptions AddEnumsWithValuesFixFilters(this SwaggerGenOptions swaggerGenOptions, bool includeDescriptionFromAttribute = false)
         {
             swaggerGenOptions.SchemaFilter<XEnumNamesSchemaFilter>(includeDescriptionFromAttribute);
             swaggerGenOptions.ParameterFilter<XEnumNamesParameterFilter>(includeDescriptionFromAttribute);
             swaggerGenOptions.DocumentFilter<DisplayEnumsWithValuesDocumentFilter>(includeDescriptionFromAttribute);
+            return swaggerGenOptions;
         }
 
-        internal static string GetFieldAttributeDescription(this Type enumType, object enumField, int attributeNumber)
-        {
-            if (!enumType.IsEnum)
-                return string.Empty;
-            var memInfo = enumType.GetMember(enumField.ToString());
-            var attributes = memInfo[0].GetCustomAttributes(typeof(DescriptionAttribute), false);
-            if (attributes.Length > 0)
-                return (attributes[attributeNumber] as DescriptionAttribute)?.Description ?? string.Empty;
-            return string.Empty;
-        }
+        #endregion
     }
 }