+++ test/WebApi3.1-Swashbuckle/Controllers/SecondController.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
+
+namespace WebApi3._1_Swashbuckle.Controllers
+{
+    /// <summary>
+    /// Second controller.
+    /// </summary>
+    [Produces("application/json")]
+    [Route("api/[controller]")]
+    [ApiController]
+    [SwaggerTag("Second controller")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(420)]
+    public class SecondController : ControllerBase
+    {
+        /// <summary>
+        /// Test action.
+        /// </summary>
+        [HttpGet()]
+        public IActionResult TestAction()
+        {
+            return NoContent();
+        }
+    }
+}